[FEAT] 회원 프로젝트 조회 기능 구현
